Moving namespaces around to help with doc generation
