+++ examples/Cli/Delegates/Program.cs
@@ -1,6 +1,7 @@
+using Spectre.Console;
 using Spectre.Console.Cli;
 
-namespace Spectre.Console.Examples;
+namespace Delegates;
 
 public static partial class Program
 {
