+++ examples/Cli/Injection/IGreeter.cs
@@ -1,4 +1,6 @@
-namespace Spectre.Console.Examples;
+using Spectre.Console;
+
+namespace Injection;
 
 public interface IGreeter
 {
