+++ examples/Cli/Injection/Infrastructure/TypeResolver.cs
@@ -1,8 +1,7 @@
 using System;
-using Microsoft.Extensions.DependencyInjection;
 using Spectre.Console.Cli;
 
-namespace Spectre.Console.Examples;
+namespace Injection.Infrastructure;
 
 public sealed class TypeResolver : ITypeResolver, IDisposable
 {
