+++ examples/Cli/Logging/Infrastructure/TypeRegistrar.cs
@@ -2,7 +2,7 @@ using System;
 using Microsoft.Extensions.DependencyInjection;
 using Spectre.Console.Cli;
 
-namespace Spectre.Console.Examples;
+namespace Logging.Infrastructure;
 
 public sealed class TypeRegistrar : ITypeRegistrar
 {
