+++ examples/Console/Emojis/Program.cs
@@ -1,4 +1,6 @@
-namespace Spectre.Console.Examples;
+using Spectre.Console;
+
+namespace Emojis;
 
 public static class Program
 {
