+++ examples/Console/Figlet/Program.cs
@@ -1,4 +1,6 @@
-namespace Spectre.Console.Examples;
+using Spectre.Console;
+
+namespace Figlet;
 
 public static class Program
 {
