+++ examples/Console/Info/Program.cs
@@ -1,4 +1,6 @@
-namespace Spectre.Console.Examples;
+using Spectre.Console;
+
+namespace Info;
 
 public static class Program
 {
