+++ examples/Console/Links/Program.cs
@@ -1,4 +1,6 @@
-namespace Spectre.Console.Examples;
+using Spectre.Console;
+
+namespace Links;
 
 public static class Program
 {
