+++ examples/Console/Rules/Program.cs
@@ -1,4 +1,6 @@
-namespace Spectre.Console.Examples;
+using Spectre.Console;
+
+namespace Rules;
 
 public static class Program
 {
