+++ examples/Console/Trees/Program.cs
@@ -1,4 +1,6 @@
-namespace Spectre.Console.Examples;
+using Spectre.Console;
+
+namespace Trees;
 
 public static class Program
 {
